refactor: Make the project structure more Pythonic :snake:
