+++ README.md
@@ -9,5 +9,5 @@ https://github.com/dudaduarte/guess_what.git
 
 Agora, inicie o menu dos jogos com o comando abaixo e siga as instruções que aparecerão no terminal :)
 ```sh
-python main.py
+python -m guess_what
 ```
