+++ guess_what/guessing.py
